Changing what kind of resources are fetched
- the node labels should contain at least one of the selected filters
- the node labels should contain the `DataResource` label too
- only main nodes (uri which does not start with `pnode://`) are returned as result.

Code cleanup: Resource, ResourceInput, ServiceOffering, ServiceOfferingInput were simplified and moved to the correct files.
