+++ src/types/resources.model.ts
@@ -1,4 +1,7 @@
-export interface ResourceType {
-    id: string,
-    label: string
+export interface Resource {
+    labels: string[],
+    name: string,
+    description: string,
+    uri: string,
+    claimsGraphUri: string,
 }
